refactor: 패키지 구조 변경
strategy, enumeration 패키지 추가
