+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.enumeration.LottoResult;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
