+++ src/main/java/lotto/strategy/NumberSelectStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.strategy;
 
 import java.util.List;
