+++ src/test/java/lotto/LottoGameTest.java
@@ -2,7 +2,7 @@ package lotto;
 
 import java.util.stream.Stream;
 import lotto.domain.LottoGame;
-import lotto.domain.RandomNumberSelectStrategy;
+import lotto.strategy.RandomNumberSelectStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
